Fixed card flickering during selection
Added new drawing functions to the Player class that are separate from
the Card draw functions. Be warned: the code is horrible and you should
not look at it. I also added a rank_char function to the Card class that
gets the correct character for the given Card, which has the potential
to be more useful in certain situations than the raw enum the Card uses
to store its rank.